rename content dialog to modal dialog
