+++ packages/modal-dialog/src/Modal-Dialog.stories.tsx
@@ -2,7 +2,7 @@ import { Button } from '@spark-web/button';
 import { Text } from '@spark-web/text';
 import type { ComponentMeta, ComponentStory } from '@storybook/react';
 
-import { ContentDialog } from './content-dialog';
+import { ContentDialog } from './Modal-Dialog';
 
 export default {
   title: 'Feedback & Overlays / Content Dialog',
